Utility PositionConjugacyClass
These files contains the function PositionConjugacyClass(G, x) for
identifying the number of the G-conjugacy class of the element x in
G.

This uses a tree structure to compute class invariants until
the class is determined, see ConjugacyClassInvariants(G).

In this version only very basic invariants are used, e.g.
cycle types for permutations, characteristic and minimal polynomial for
matrices, and as fallback always conjugation tests against candidates.

This could be enhanced with more functions that find other invariants.