combine `IrrUnger` with `IrrBaumClausen` (#16)
* combine `IrrUnger` with `IrrBaumClausen`

In the `Irr` method that calls `IrrUnger`,
if the group is abelian-by-supersolvable, call `IrrBaumClausen`,
set the irreducibles, and return the result.

(For other groups, the result of `IrrBaumClausen` would be incomplete.
In this situation, using the partial list of irreducibles in the
initialization of `IrrUnger` looks complicated.
We do not try to do this here.)

* add tests

in order to make Codecov happy